+++ src/main/java/com/newer/sc/manager/mapper/MteacherMapper.java
@@ -1,6 +1,5 @@
 package com.newer.sc.manager.mapper;
 
-import java.math.BigInteger;
 import java.util.List;
 
 
@@ -43,7 +42,7 @@ public interface MteacherMapper {
 	@Results(id="teacherMap",value = {
 			@Result(column = "t_coid",property = "college",javaType =College.class,one = @One(select = "com.newer.sc.manager.mapper.McollegeMapper.findById")),
 			})
-	Teacher findById(BigInteger id);
+	Teacher findById(int id);
 	
 	/**
 	 * 查询所有老师信息
